+++ src/Queries/Contracts/Paginatable.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Components\Queries\Contracts;
+
+interface Paginatable
+{
+    public function page(): int;
+
+    public function perPage(): int;
+}